requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
bcrypt
python-multipart
